+++ packages/snaps-controllers/src/snaps/SnapController.test.tsx
@@ -1,7 +1,4 @@
-import {
-  deriveStateFromMetadata,
-  getPersistentState,
-} from '@metamask/base-controller';
+import { deriveStateFromMetadata } from '@metamask/base-controller';
 import { encrypt } from '@metamask/browser-passworder';
 import {
   createAsyncMiddleware,
@@ -88,7 +85,10 @@ import {
   STATE_DEBOUNCE_TIMEOUT,
 } from './constants';
 import { SnapsRegistryStatus } from './registry';
-import type { SnapControllerState } from './SnapController';
+import type {
+  PersistedSnapControllerState,
+  SnapControllerState,
+} from './SnapController';
 import {
   SNAP_APPROVAL_INSTALL,
   SNAP_APPROVAL_RESULT,
@@ -2854,7 +2854,7 @@ describe('SnapController', () => {
         },
       });
 
-      expect(rootMessenger.call).toHaveBeenNthCalledWith(
+      expect(messenger.call).toHaveBeenNthCalledWith(
         4,
         'ExecutionService:handleRpcRequest',
         MOCK_SNAP_ID,
@@ -5184,8 +5184,8 @@ describe('SnapController', () => {
         },
       });
 
-      expect(rootMessenger.call).toHaveBeenCalledTimes(5);
-      expect(rootMessenger.call).toHaveBeenCalledWith(
+      expect(options.messenger.call).toHaveBeenCalledTimes(5);
+      expect(options.messenger.call).toHaveBeenCalledWith(
         'ExecutionService:handleRpcRequest',
         MOCK_SNAP_ID,
         {
@@ -6314,7 +6314,7 @@ describe('SnapController', () => {
       });
       const [snapController] = getSnapControllerWithEES(snapControllerOptions);
 
-      expect(rootMessenger.call).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.call).toHaveBeenCalledWith(
         'PermissionController:grantPermissions',
         {
           approvedPermissions: {
@@ -6330,7 +6330,7 @@ describe('SnapController', () => {
         },
       );
 
-      expect(rootMessenger.publish).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.publish).toHaveBeenCalledWith(
         'SnapController:snapInstalled',
         getTruncatedSnap(),
         METAMASK_ORIGIN,
@@ -6414,12 +6414,12 @@ describe('SnapController', () => {
         () => MOCK_SNAP_PERMISSIONS,
       );
 
-      expect(rootMessenger.call).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.call).toHaveBeenCalledWith(
         'PermissionController:grantPermissions',
         { approvedPermissions, subject: { origin: 'npm:filsnap' } },
       );
 
-      expect(rootMessenger.call).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.call).toHaveBeenCalledWith(
         'PermissionController:grantPermissions',
         {
           approvedPermissions,
@@ -6463,7 +6463,7 @@ describe('SnapController', () => {
       });
       const [snapController] = getSnapControllerWithEES(snapControllerOptions);
 
-      expect(rootMessenger.call).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.call).toHaveBeenCalledWith(
         'PermissionController:grantPermissions',
         {
           approvedPermissions: {
@@ -6543,14 +6543,14 @@ describe('SnapController', () => {
       });
       const [snapController] = getSnapControllerWithEES(snapControllerOptions);
 
-      expect(rootMessenger.call).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.call).toHaveBeenCalledWith(
         'PermissionController:revokePermissions',
         {
           [MOCK_SNAP_ID]: ['snap_dialog'],
         },
       );
 
-      expect(rootMessenger.call).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.call).toHaveBeenCalledWith(
         'PermissionController:grantPermissions',
         {
           approvedPermissions: {
@@ -6561,7 +6561,7 @@ describe('SnapController', () => {
         },
       );
 
-      expect(rootMessenger.publish).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.publish).toHaveBeenCalledWith(
         'SnapController:snapUpdated',
         getTruncatedSnap({
           version: '1.2.3',
@@ -6630,7 +6630,7 @@ describe('SnapController', () => {
       });
       const [snapController] = getSnapControllerWithEES(snapControllerOptions);
 
-      expect(rootMessenger.call).toHaveBeenCalledTimes(0);
+      expect(snapControllerOptions.messenger.call).toHaveBeenCalledTimes(0);
 
       snapController.destroy();
     });
@@ -6680,7 +6680,7 @@ describe('SnapController', () => {
       });
       const [snapController] = getSnapControllerWithEES(snapControllerOptions);
 
-      expect(rootMessenger.call).toHaveBeenCalledWith(
+      expect(snapControllerOptions.messenger.call).toHaveBeenCalledWith(
         'PermissionController:grantPermissions',
         {
           approvedPermissions: {
@@ -6910,7 +6910,7 @@ describe('SnapController', () => {
 
       await new Promise((resolve) => setTimeout(resolve, 10));
 
-      expect(rootMessenger.call).toHaveBeenNthCalledWith(
+      expect(messenger.call).toHaveBeenNthCalledWith(
         6,
         'ExecutionService:handleRpcRequest',
         MOCK_SNAP_ID,
@@ -6976,7 +6976,7 @@ describe('SnapController', () => {
 
       await new Promise((resolve) => setTimeout(resolve, 10));
 
-      expect(rootMessenger.call).toHaveBeenNthCalledWith(
+      expect(messenger.call).toHaveBeenNthCalledWith(
         6,
         'ExecutionService:handleRpcRequest',
         MOCK_SNAP_ID,
@@ -10172,12 +10172,13 @@ describe('SnapController', () => {
       expect(updatedSnap.version).toStrictEqual(updateVersion);
       expect(updatedSnap.preinstalled).toBe(true);
 
-      expect(rootMessenger.call).toHaveBeenNthCalledWith(
+      expect(messenger.call).toHaveBeenNthCalledWith(
         7,
         'PermissionController:revokePermissions',
         { [snapId]: [SnapEndowments.Rpc, SnapEndowments.LifecycleHooks] },
       );
-      expect(rootMessenger.call).toHaveBeenNthCalledWith(
+
+      expect(messenger.call).toHaveBeenNthCalledWith(
         8,
         'PermissionController:grantPermissions',
         {
@@ -12621,7 +12622,7 @@ describe('SnapController', () => {
         deriveStateFromMetadata(
           controller.state,
           controller.metadata,
-          'anonymous',
+          'includeInDebugSnapshot',
         ),
       ).toMatchInlineSnapshot(`{}`);
     });
@@ -12747,15 +12748,16 @@ describe('SnapController', () => {
         );
 
         // persist the state somewhere
-        const persistedState = getPersistentState<SnapControllerState>(
+        const persistedState = deriveStateFromMetadata<SnapControllerState>(
           firstSnapController.state,
           firstSnapController.metadata,
+          'persist',
         );
 
         // create a new controller
         const secondSnapController = getSnapController(
           getSnapControllerOptions({
-            state: persistedState,
+            state: persistedState as PersistedSnapControllerState,
           }),
         );
 
@@ -12786,15 +12788,16 @@ describe('SnapController', () => {
         expect(firstSnapController.state.snaps[MOCK_SNAP_ID]).toBeDefined();
 
         // persist the state somewhere
-        const persistedState = getPersistentState<SnapControllerState>(
+        const persistedState = deriveStateFromMetadata<SnapControllerState>(
           firstSnapController.state,
           firstSnapController.metadata,
+          'persist',
         );
 
         // create a new controller
         const secondSnapController = getSnapController(
           getSnapControllerOptions({
-            state: persistedState,
+            state: persistedState as PersistedSnapControllerState,
           }),
         );