+++ packages/snaps-controllers/src/test-utils/execution-environment.ts
@@ -1,5 +1,6 @@
 import { JsonRpcEngine } from '@metamask/json-rpc-engine';
 import { createEngineStream } from '@metamask/json-rpc-middleware-stream';
+import { Messenger } from '@metamask/messenger';
 import { logError, type SnapRpcHookArgs } from '@metamask/snaps-utils';
 import type { MockControllerMessenger } from '@metamask/snaps-utils/test-utils';
 import { pipeline } from 'readable-stream';
@@ -19,21 +20,21 @@ export const getNodeEESMessenger = (
     ExecutionServiceActions,
     ExecutionServiceEvents
   >,
-) =>
-  messenger.getRestricted({
-    name: 'ExecutionService',
-    allowedEvents: [
-      'ExecutionService:unhandledError',
-      'ExecutionService:outboundRequest',
-      'ExecutionService:outboundResponse',
-    ],
-    allowedActions: [
-      'ExecutionService:executeSnap',
-      'ExecutionService:handleRpcRequest',
-      'ExecutionService:terminateAllSnaps',
-      'ExecutionService:terminateSnap',
-    ],
-  });
+) => {
+  const executionServiceMessenger = new Messenger<
+    'ExecutionService',
+    ExecutionServiceActions,
+    ExecutionServiceEvents,
+    any
+  >({ namespace: 'ExecutionService', parent: messenger });
+
+  messenger.unregisterActionHandler('ExecutionService:handleRpcRequest');
+  messenger.unregisterActionHandler('ExecutionService:executeSnap');
+  messenger.unregisterActionHandler('ExecutionService:terminateSnap');
+  messenger.unregisterActionHandler('ExecutionService:terminateAllSnaps');
+
+  return executionServiceMessenger;
+};
 
 export const getNodeEES = (
   messenger: ReturnType<typeof getNodeEESMessenger>,
@@ -64,6 +65,10 @@ export const getNodeEES = (
   });
 
 export class ExecutionEnvironmentStub implements ExecutionService {
+  name: 'ExecutionService' = 'ExecutionService' as const;
+
+  state = null;
+
   constructor(messenger: ReturnType<typeof getNodeEESMessenger>) {
     messenger.registerActionHandler(
       `ExecutionService:handleRpcRequest`,